refactor(activity): update pending transaction action visibility logic
